+++ CHANGES.md
@@ -2,6 +2,7 @@
   * Save clipboard to layout
   * RSAM plot screen capture
   * Use S-picks in location algorithm
+  * Fix Hypo71.config read 
 
 ## Version 2.8.1
   * Upgraded seedlink library